fix: make priority select a component and remove unnecessary code
